Major restructuring
Really worked to clean things. I know handle the thinking and content pieces better by instantiating gr.ChatMessage which I edit appropriately as we stream back chunks from the LLM. I also only send back necessary information.

I also learned how to use the "metadata" field in the gr.ChatMessage to more easily and in a much cleaner way to show the user that the LLM is "thinking".

build_message_history now keeps out any message in the history that has a "metadata" field so that thinking and tool calling don't junk up the messages to send to the LLM.